Remove directories
This is me being lazy, but I couldn't figure out how to make images work
for both the compiled version and in github, since one uses root as a
reference while the other the directory you are in. Making everything
live in the root is good enough. The alternative would require something
different of bash, or someone who knows bash better than me.